+++ lib/kag/bot/plugin.rb
@@ -18,8 +18,10 @@ module KAG
         end
       end
 
-      match "restart", :method => :evt_restart
-      def evt_restart(m)
+      command :restart,{},
+        summary: "Restart the bot",
+        admin: true
+      def restart(m)
         if is_admin(m.user)
           cmd = (KAG::Config.instance[:restart_method] or "nohup sh gather.sh &")
           debug cmd
@@ -29,8 +31,9 @@ module KAG
         end
       end
 
-      match /is_admin (.+)/, :method => :evt_am_i_admin
-      def evt_am_i_admin(m,nick)
+      command :is_an_admin,{nick: :string},
+        summary: "See if the specified user is an Admin"
+      def is_an_admin(m,nick)
         u = User(nick)
         if is_admin(u)
           reply m,"Yes, #{nick} is an admin!"
@@ -39,23 +42,15 @@ module KAG
         end
       end
 
-      match "reload_config", :method => :evt_reload_config
-      def evt_reload_config(m)
+      command :reload_config,{},
+        summary: "Reload the configuration file",
+        admin: true
+      def reload_config(m)
         if is_admin(m.user)
           KAG::Config.instance.reload
           m.reply "Configuration reloaded."
         end
       end
-=begin
-      match "help", :method => :evt_help
-      def evt_help(m)
-        unless is_banned?(m.user)
-          msg = "Commands: !add, !rem, !list, !status, !help, !end, !report, !reports [nick]"
-          msg = msg + ", !rem [nick], !add [nick], !add_silent, !rem_silent, !unreport, !clear, !restart, !quit" if is_admin(m.user)
-          User(m.user.nick).send(msg)
-        end
-      end
-=end
     end
   end
 end